Version 1.20.3-at.20211001.01. Fix implicit conversion off_t to int.

+++ src/warc.c
@@ -419,7 +419,7 @@ static bool
 warc_write_block_from_file (FILE *data_in)
 {
   /* Add the Content-Length header. */
-  int content_length_i;
+  off_t content_length_i;
   char content_length[MAX_INT_TO_STRING_LEN(off_t)];
   size_t buffer_size;
 #ifdef HAVE_ZSTD
